NEXT-18351 - Change npm install to clean-install for plugins fixes #121

+++ bin/build-administration.sh
@@ -31,7 +31,11 @@ if [[ $(command -v jq) ]]; then
         if [[ -f "$path/package.json" && ! -f "$path/node_modules" && $name != "administration" ]]; then
             echo "=> Installing npm dependencies for ${name}"
 
-            npm install --prefix "$path"
+            if [[ -f "$path/package-lock.json" ]]; then
+                npm clean-install --prefix "$path"
+            else
+                npm install --prefix "$path"
+            fi
         fi
     done
     cd "$OLDPWD" || exit
